pyproject.toml
[project]
name = "openstack_mcp_server"
version = "0.1.0"
description = "Openstack MCP server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.9.0",
    "openstacksdk>=4.5.0"
]

[project.optional-dependencies]
dev = [
    "tox==4.25.0",
    "pylint==3.0.0",
    "ruff==0.9.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]

[project.scripts]
openstack_mcp_server = "server.server:main"
